use a shared trpc server context between TT and PT
