+++ apps/protailwind/src/server/routers/_app.ts
@@ -1,21 +1,15 @@
-/**
- * This file contains the root router of your tRPC-backend
- */
-import {progressRouter} from './progress'
-import {convertkitRouter} from './convertkit'
-import {router} from '../trpc'
+import {
+  mergeRouters,
+  router,
+  skillLessonRouter,
+} from '@skillrecordings/skill-lesson'
 import {lessonResourcesRouter} from './resources'
 
-/**
- * Create your application's root router
- * If you want to use SSG, you need export this
- * @link https://trpc.io/docs/ssg
- * @link https://trpc.io/docs/router
- */
-export const appRouter = router({
-  progress: progressRouter,
-  convertkit: convertkitRouter,
-  resources: lessonResourcesRouter,
-})
+export const appRouter = mergeRouters(
+  router({
+    resources: lessonResourcesRouter,
+  }),
+  skillLessonRouter,
+)
 
 export type AppRouter = typeof appRouter
