+++ apps/skill-front-support/src/utils/trpc.ts
@@ -1,4 +1,3 @@
-// src/utils/trpc.ts
 import {httpBatchLink, loggerLink} from '@trpc/client'
 import {createTRPCNext} from '@trpc/next'
 import type {AppRouter} from '../server/trpc/router/_app'
