+++ apps/total-typescript/src/pages/api/trpc/[trpc].ts
@@ -2,7 +2,7 @@
  * This file contains the tRPC http response handler and context creation for Next.js
  */
 import * as trpcNext from '@trpc/server/adapters/next'
-import {createContext} from 'server/context'
+import {createContext} from '@skillrecordings/skill-lesson'
 import {AppRouter, appRouter} from 'server/routers/_app'
 
 export default trpcNext.createNextApiHandler<AppRouter>({
