+++ apps/total-typescript/src/server/routers/pricing.ts
@@ -1,4 +1,4 @@
-import {publicProcedure, router} from '../trpc'
+import {publicProcedure, router} from '@skillrecordings/skill-lesson'
 import {z} from 'zod'
 import {getSdk} from '@skillrecordings/database'
 import {
