+++ packages/skill-lesson/index.tsx
@@ -1,1 +1,4 @@
 export * from './utils/ability'
+export * from './server/routers/skill-lesson-router'
+export * from './server/context'
+export * from './server/trpc'
