+++ packages/skill-lesson/package.json
@@ -22,9 +22,12 @@
     "@skillrecordings/ability": "workspace:*",
     "@skillrecordings/analytics": "workspace:*",
     "@skillrecordings/convertkit-sdk": "workspace:*",
+    "@skillrecordings/database": "workspace:*",
     "@skillrecordings/time": "workspace:*",
     "@skillrecordings/react": "workspace:*",
     "@skillrecordings/config": "workspace:*",
+    "@trpc/server": "^10.7.0",
+    "next-auth": "=4.10.3",
     "lodash": "^4.17.21",
     "cookie": "^0.5.0",
     "@types/cookie": "^0.4.1",
@@ -43,6 +46,7 @@
     "refractor": "^4.5.0",
     "@types/speakingurl": "^13.0.3",
     "groq": "^2.15.0",
+    "superjson": "^1.11.0",
     "typescript": "4.9.4"
   }
 }
