+++ packages/skill-lesson/server/routers/skill-lesson-router.ts
@@ -1,0 +1,19 @@
+/**
+ * This file contains the root router of your tRPC-backend
+ */
+import {progressRouter} from './progress'
+import {convertkitRouter} from './convertkit'
+import {router} from '../trpc'
+
+/**
+ * Create your application's root router
+ * If you want to use SSG, you need export this
+ * @link https://trpc.io/docs/ssg
+ * @link https://trpc.io/docs/router
+ */
+export const skillLessonRouter = router({
+  progress: progressRouter,
+  convertkit: convertkitRouter,
+})
+
+export type SkillLessonRouter = typeof skillLessonRouter
