+++ packages/skill-lesson/server/trpc.ts
@@ -8,9 +8,9 @@
  * @see https://trpc.io/docs/v10/procedures
  */
 
-import {Context} from './context'
 import {initTRPC} from '@trpc/server'
 import superjson from 'superjson'
+import {Context} from './context'
 
 const t = initTRPC.context<Context>().create({
   /**
